refactor(meta/snapshot): improve performance
- Use buffered writer for snapshot writer

- Avoid alloc a temp big vec for snapshot

- Refine StoreInner::build_compacted_snapshot, move meta building to SnapshotViewV002
